feat: Relocated the view home and its components
